+++ pytest_factoryboy/__init__.py
@@ -3,9 +3,12 @@
 __version__ = '1.0.3'
 
 try:
-    from .fixture import register
+    from .fixture import register, LazyFixture
 
-    __all__ = [register.__name__]
+    __all__ = [
+        register.__name__,
+        LazyFixture.__name__,
+    ]
 except ImportError:  # pragma: no cover
     # avoid import errors when only __version__ is needed (for setup.py)
     pass